requirements.txt
python-slugify==1.1.4
requests==2.8.1
Unidecode==0.4.18
